requirements.txt
absl-py==1.3.0
aiobotocore==2.5.4
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aioitertools==0.12.0
aiosignal==1.3.1
alembic==1.14.0
altair==4.2.2
annotated-types==0.6.0
anyio==3.7.1
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.5.2
asn1crypto==1.5.1
astor==0.8.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
atpublic==5.0
attrdict==2.0.1
attrs==24.2.0
Babel==2.12.1
backoff==2.2.1
bce-python-sdk==0.8.90
bcrypt==4.2.1
beautifulsoup4==4.10.0
bleach==6.2.0
blinker==1.6.2
blosc2==2.0.0
botocore==1.31.17
branca==0.7.1
bs4==0.0.1
build==1.2.2.post1
cachetools==5.2.0
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==2.0.12
chroma-hnswlib==0.7.6
chromadb==0.5.20
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
cryptography==43.0.3
cssselect==1.2.0
cssselect2==0.7.0
cssutils==2.7.1
Cython==3.0.0
defusedxml==0.7.1
Deprecated==1.2.15
diff-match-patch==20200713
dill==0.3.9
distlib==0.3.8
distro==1.8.0
Django==2.2.28
django-import-export==2.9.0
dukpy==0.3.0
durationpy==0.9
einops==0.4.1
elementpath==4.1.5
esprima==4.0.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fairscale==0.4.8
faiss-cpu==1.7.2
faiss-gpu==1.7.2
fastapi==0.115.5
fastjsonschema==2.21.1
ffpyplayer==4.5.0
filelock==3.16.1
fire==0.5.0
Flask==2.3.3
flatbuffers==23.5.9
folium==0.16.0
fqdn==1.5.1
freetype-py==2.4.0
frozenlist==1.5.0
fsspec==2023.9.0
gast==0.4.0
gensim==4.3.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==2.0.2
grpcio==1.68.1
gym==0.25.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.8.0
html2text==2020.1.16
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
imageio-ffmpeg==0.4.8
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
javascripthon==0.12
jax==0.4.10
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.2.0
json-tricks==3.17.3
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.6.4
keras==2.12.0
kubernetes==31.0.0
labml==0.4.159
labml-helpers==0.4.89
labml-nn==0.4.129
langchain==0.3.9
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langsmith==0.1.147
libclang==16.0.0
lmdb==1.4.1
lxml==4.9.3
Mako==1.3.6
Markdown==3.4.1
markdown-it-py==3.0.0
markdownify==0.11.6
MarkupPy==1.14
MarkupSafe==2.1.2
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.1.0
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack-numpy==0.4.8
multidict==6.1.0
munch==2.5.0
narwhals==1.14.2
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numexpr==2.8.5
numpy==1.23.2
nvidia-nccl-cu12==2.23.4
oauthlib==3.2.2
odfpy==1.4.1
onnxruntime==1.20.1
openai==1.0.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
opt-einsum==3.3.0
orjson==3.10.12
outcome==1.3.0.post0
overrides==7.7.0
paddle-bfloat==0.1.7
pandas==2.2.3
pandocfilters==1.5.1
pdf2docx==0.5.6
pillow==11.0.0
platformdirs==2.6.2
plotly==5.24.1
posthog==3.7.4
premailer==3.10.0
prometheus_client==0.21.0
propcache==0.2.1
protobuf==5.28.3
psychtoolbox==3.0.19.0
psycopg2-binary==2.9.6
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyChatGPT==0.4.3.3
pyclipper==1.3.0.post4
pycparser==2.22
pycryptodome==3.18.0
pydantic==2.10.1
pydantic_core==2.27.1
pydeck==0.9.1
pyglet==1.5.27
PyJWT==2.10.0
PyMuPDF==1.20.2
pyOpenSSL==24.2.1
pyparallel==0.2.2
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQt6-Qt6==6.5.2
PyQt6-sip==13.5.2
pyserial==3.5
PySocks==1.7.1
python-bidi==0.4.2
python-decouple==3.8
python-docx==0.8.11
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.2
python-vlc==3.0.18122
python-xlib==0.33
pytimeparse==1.1.8
pytz==2023.3
PyVirtualDisplay==3.0
PyYAML==6.0.2
rapidfuzz==3.2.0
rarfile==4.0
referencing==0.35.1
regex==2023.5.5
reportlab==3.6.13
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
s3fs==2023.9.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.15.2
Send2Trash==1.8.3
sentence-transformers==3.3.1
shapely==2.0.1
shellingham==1.5.4
six==1.16.0
sklearn==0.0.post5
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.1
snowflake==0.8.0
snowflake-connector-python==3.12.3
snowflake-ml-python==1.7.2
snowflake-snowpark-python==1.25.0
snowflake._legacy==1.0.0
snowflake.core==0.8.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.4.3
starlette==0.41.3
streamlit==1.40.1
streamlit-aggrid==1.0.5
svglib==1.4.1
sympy==1.13.3
tablib==3.2.1
tenacity==9.0.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.4.0
tls-client==0.2.2
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
torch==1.12.1
torchtext==0.13.1
torchvision==0.13.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
trash-cli==0.23.2.13.2
trio==0.23.1
trio-websocket==0.11.1
trulens==1.2.9
trulens-core==1.2.9
trulens-dashboard==1.2.9
trulens-feedback==1.2.9
trulens-providers-cortex==1.2.9
trulens_eval==1.2.9
typer==0.14.0
types-python-dateutil==2.9.0.20241003
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
undetected-chromedriver==3.5.3
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.32.1
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==2.3.7
wget==3.2
widgetsnbextension==4.0.13
wrapt==1.14.1
wsproto==1.2.0
xgboost==2.1.3
xlrd==1.2.0
xlwt==1.3.0
xmlschema==2.4.0
xyzservices==2023.10.1
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==6.0
